Automation of Algoritm
I have successfully created a script to continuously run the data collection script as well as the algorithm script. I utilized the usage of pathlib as well as python logging to provide substantial feedback on the code. The anomaly detection code has been further enhanced to check the numerous data types that we are collecting. Currently it does not trigger any events based on the anomaly result.